Add support for multiple/all project queries
Changed the handling of the :project parameter so that it will only be applied
as a --project command line flag when there is a single project given.

If no project is provided as a parameter, the query will instead be wrapped in
'project IS NOT EMPTY AND (...)', which is how with JiraCLI you have to do it to
query across all projects.

If more than one project is given, as a comma separated list, the query is
instead wrapped in 'project IN (...) and (...)'.